requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.27
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.7
openai>=1.12.0
python-dotenv>=1.0.0
alembic>=1.13.1
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dateutil>=2.8.2
email-validator>=2.1.0
bcrypt>=4.0.1
